backend/requirements.txt
Flask>=2.0
python-telegram-bot>=13.0
numpy>=1.21
scikit-learn>=1.0
